# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pharmacy API",
        "description": "RESTful API theo chuẩn JSON:API",
        "contact": {
            "email": "dongockhanh2003.utt@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Main Server"
        }
    ],
    "paths": {
...
        "/v1/admin/medicines": {
            "get": {
                "tags": [
                    "Medicines"
                ],
                "summary": "Lấy danh sách thuốc",
                "description": "Truy xuất tất cả các thuốc cùng với thông tin danh mục và nhà cung cấp liên quan",
                "operationId": "listMedicine",
                "responses": {
                    "200": {
                        "description": "Danh sách thuốc",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Danh sách thuốc"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy thuốc",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "null"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Không tìm thấy thuốc"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/v1/admin/medicines/delete/{id}": {
            "delete": {
                "tags": [
                    "Medicines"
                ],
                "summary": "Xóa một thuốc",
                "description": "Xóa một thuốc theo ID",
                "operationId": "deleteMedicine",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của thuốc cần xóa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thuốc thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đã xóa thuốc thành công"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy thuốc",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Không tìm thấy thuốc"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Xác thực bằng JWT Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}